ref: Move scoping back to relay-quotas
