+++ relay-server/src/extractors/request_meta.rs
@@ -15,7 +15,7 @@ use relay_common::{
     Auth, Dsn, ParseAuthError, ParseDsnError, ParseProjectIdError, ParseProjectKeyError, ProjectId,
     ProjectKey, Scheme,
 };
-use relay_project_config::quota::Scoping;
+use relay_quotas::Scoping;
 
 use crate::body::PeekLine;
 use crate::extractors::ForwardedFor;
