+++ relay-server/src/utils/metrics_rate_limits.rs
@@ -3,8 +3,8 @@ use chrono::Utc;
 
 use relay_common::{DataCategory, UnixTimestamp};
 use relay_metrics::{MetricNamespace, MetricResourceIdentifier, MetricsContainer};
-use relay_project_config::quota::{ItemScoping, Quota, Scoping};
-use relay_quotas::RateLimits;
+use relay_project_config::quota::Quota;
+use relay_quotas::{ItemScoping, RateLimits, Scoping};
 
 use crate::actors::outcome::{DiscardReason, Outcome, TrackOutcome};
